# storage/whatsapp-api-docs/whatsapp-api-swagger-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp API",
        "description": "This is an API documentation for the WhatsApp-like chat application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chatrooms": {
            "get": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "List all chatrooms",
                "operationId": "ce7fe6f309c557b6e5c5b4fa2acbed05",
                "responses": {
                    "200": {
                        "description": "A list of chatrooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "General Chat"
                                            },
                                            "max_members_count": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-07T10:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-07T10:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "Create a new chatroom",
                "operationId": "ce9864db3beaf86b201621bda4c884da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "max_members_count"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "General Chat"
                                    },
                                    "max_members_count": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chatroom created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "General Chat"
                                        },
                                        "max_members_count": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-07T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-07T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chatrooms/{chatroomId}/join": {
            "get": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "User enters a chatroom",
                "operationId": "0df7071a468be9e1c76e73d9bcdb6c23",
                "responses": {
                    "200": {
                        "description": "User enters a chatroom."
                    },
                    "401": {
                        "description": "Invalid chatroom."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chatrooms/{chatroomId}/leave": {
            "get": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "User leaves a chatroom",
                "operationId": "bff234f59d9714cdc916ed31a82fcd60",
                "responses": {
                    "200": {
                        "description": "Left chatroom successfully."
                    },
                    "401": {
                        "description": "Invalid chatroom."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}